+++ edgy/core/db/querysets/base.py
@@ -307,17 +307,17 @@ class BaseQuerySet(
         )
 
         for key, value in cleaned_kwargs.items():
-            model_class, field_name, op, related_str, _, cross_db_remainder = crawl_relationship(
-                self.model_class, key
+            crawl_result = crawl_relationship(self.model_class, key)
+            if crawl_result.forward_path:
+                select_related.add(crawl_result.forward_path)
+            field = crawl_result.model_class.meta.fields.get(
+                crawl_result.field_name, clauses_mod.generic_field
             )
-            if related_str:
-                select_related.add(related_str)
-            field = model_class.meta.fields.get(field_name, clauses_mod.generic_field)
-            if cross_db_remainder:
+            if crawl_result.cross_db_remainder:
                 assert field is not clauses_mod.generic_field
                 fk_field = cast(BaseForeignKey, field)
                 sub_query = (
-                    fk_field.target.query.filter(**{cross_db_remainder: value})
+                    fk_field.target.query.filter(**{crawl_result.cross_db_remainder: value})
                     .only(*fk_field.related_columns.keys())
                     .values_list(fields=fk_field.related_columns.keys())
                 )
@@ -328,7 +328,7 @@ class BaseQuerySet(
                     *,
                     _field: BaseFieldType = field,
                     _sub_query: QuerySet = sub_query,
-                    _prefix: str = related_str,
+                    _prefix: str = crawl_result.forward_path,
                 ) -> Any:
                     table = tables_and_models[_prefix][0]
                     fk_tuple = sqlalchemy.tuple_(
@@ -348,9 +348,9 @@ class BaseQuerySet(
                     *,
                     _field: BaseFieldType = field,
                     _value: Any = value,
-                    _op: str | None = op,
-                    _prefix: str = related_str,
-                    _field_name: str = field_name,
+                    _op: str | None = crawl_result.operator,
+                    _prefix: str = crawl_result.forward_path,
+                    _field_name: str = crawl_result.field_name,
                 ) -> Any:
                     _value = await clauses_mod.parse_clause_arg(
                         _value, queryset, tables_and_models
@@ -397,15 +397,18 @@ class BaseQuerySet(
         """
         # retrieve the cache from queryset, use cache_name to identify
         cache_weak: set[str] = getattr(self, cache_name)
+        # use cache_name to load presets, this allows rapid changes in case of different
+        # required cache behaviour and validates the cache name
         match cache_name:
             case "_select_related_embedding":
                 related_element_fn: Callable[[str], str] = lambda field_name: (
-                    crawl_relationship(
+                    clauses_mod.clean_path_to_crawl_result(
                         self.model_class,
                         field_name,
                         model_database=self.database,
-                        traverse_last=True,
-                        no_operator=True,
+                        embed_parent=self.embed_parent_filters,
+                        # we need this as we have no field name here
+                        path_to_field=False,
                     ).forward_path
                 )
             case "_select_related_g_and_o":