+++ docs/release-notes.md
@@ -21,6 +21,7 @@
 - `crawl_relationship` has now a better error when a field does not exist.
 - Move `get_table_key_or_name` to `edgy.core.utils.db`.
 - Add embedded path implicit to select_related paths.
+- Unpacking and accessing `CrawlResult` (the result of `crawl_relationship`) is now deprecated. Use the results attributes instead.
 
 ### Fixed